requirements.txt
Appium-Python-Client
pytest
pytest-xdist
pytest-html
selenium
faker